Implement a function that creates a grid of tiles
The best way to be able to create a grid from a specified value was by
using a for loop to iterate through a range of values up to the
specified value and create a new tile in the grid container at every
iteration.

This is why I chose to create a new function that populates the grid
with tiles using a for loop.