Add change resolution button
The change resolution button was added with the html file and styled
properly in the css file so it appears at the top of the page above
the grid. The button was created to eventually act as a click event
listener to show a prompt for the user to enter a custom grid size.