Changed the tile hover event listener behavior
Just changing the tile color from white to black on mouse hover is
boring so the event listener was changed to set the color of the tile
to a random color.

The lightness of the tile color also turns black if it's been hovered over
more than 10 times. The tile color transitions back into full
color slowly over 10 hovers as well.